Fix motion popups sometimes getting stuck (#157)
If the popup transitioned from visible to invisible before the
status reached 'motion', the popup would get stuck in a visible
state. This changes the code to immediately set status to 'stable'
during a transition from visible to invisible if the status is
something before 'motion'.

Also fixes an issue where an animation frame could change the status
to something invalid after a transition from visible to invisible.